replace direct core lib usage with SciChartReact in drawbehindaxes
